restructure scraper to prep for hours
